修改一下version的sync，减少过多繁琐逻辑+自动同步，交给程序才是最好的~
